Renamed constants to settings
[RENAME] Constants file to settings.
